pyproject.toml
[tool.poetry]
name = "schizobench"
version = "2.0.0"
description = "Automated benchmark for evaluating LLM propensity to enable magical thinking"
authors = ["SchizoBench Team"]
readme = "README.md"
license = "MIT"
repository = "https://github.com/yourusername/schizobench"
keywords = ["ai-safety", "llm", "benchmark", "evaluation", "mental-health"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

[tool.poetry.dependencies]
python = "^3.9"
anthropic = "^0.39.0"
openai = "^1.0.0"
google-generativeai = "^0.3.0"
python-dotenv = "^1.0.0"
requests = "^2.31.0"
httpx = "0.27.2"
matplotlib = "^3.7.0"
seaborn = "^0.12.0"
pandas = "^2.0.0"
numpy = "^1.24.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.0.0"
mypy = "^1.5.0"
flake8 = "^6.1.0"
isort = "^5.12.0"
pre-commit = "^3.5.0"

[tool.poetry.scripts]
schizobench = "schizobench.cli:main"
schizobench-v2 = "run_benchmark_v2:main"
verify-setup = "verify_setup:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "tests",
]